mediscanAI/backend/VideoAssistant/requirements.txt


jupyter
ffmpeg-python
python-dotenv
shiny
git+https://github.com/jcheng5/shinymedia
pandas

matplotlib
groq
gTTS
pyttsx3
opencv-python
flask
flask-socketio
eventlet
requests
flask
flask-cors
flask-socketio
eventlet
requests
flask
flask-cors
werkzeug
opencv-python
numpy
astrapy



Flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0


requests==2.31.0
gunicorn==21.2.0
numpy>=1.24,<2.3
scipy>=1.10.0,<1.14.0




Flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0


numpy>=1.24,<2.3

# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# LangChain and AI
langchain==0.1.0
langchain-openai==0.0.5
langchain-core==0.1.0
openai==1.12.0

# Computer Vision and Image Processing
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
uuid==1.30

# Optional: For enhanced video processing
moviepy==1.0.3

# Optional: For database support (if needed later)
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1

# Optional: For better logging
colorlog==6.8.2

# Optional: For API documentation
flask-restx==1.3.0

